refactor: make folders for guards and strategy
