+++ src/auth/auth.controller.ts
@@ -8,10 +8,12 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
+
 import { Request } from 'express';
 import { CreateUserDto } from 'src/users/dto/create-user.dto';
 import { LoginUserDto } from './dto/LoginUser.dto';
 import { AuthService } from './auth.service';
+import { RefreshTokenGuard } from './guards/refresh-auth.guard';
 
 @Controller('auth')
 export class AuthController {
@@ -29,10 +31,12 @@ export class AuthController {
     return await this.authService.login(loginUser);
   }
 
-  @UseGuards(AuthGuard('jwt-refresh'))
+  @UseGuards(RefreshTokenGuard)
   @Post('refresh')
   @HttpCode(200)
   async refresh(@Req() req: Request) {
+    const user = req.user
+    console.log(user)
     return await this.authService.refresh(req.body);
   }
 }
